app/requirements.txt
Flask
psycopg2-binary
redis
PyJWT 
Flask-JWT-Extended